# api/openapi.yaml
openapi: 3.0.3
info:
  title: TradingCopilot API
  version: 0.1.0
  description: JSON:API contract target for the Go modular monolith migration.
servers:
  - url: /api
security:
  - bearerAuth: []
paths:
...
  /meetings:
    get:
      operationId: getMeetings
      parameters:
        - $ref: "#/components/parameters/ResearchTeamId"
        - $ref: "#/components/parameters/Status"
        - $ref: "#/components/parameters/TriggerSource"
        - $ref: "#/components/parameters/Tag"
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
      responses:
        "200": { $ref: "#/components/responses/MeetingCollectionDocument" }
    post:
      operationId: postMeetings
      requestBody: { $ref: "#/components/requestBodies/MeetingStartRequest" }
      responses:
        "200": { $ref: "#/components/responses/MeetingDocument" }
  /meetings/{meetingId}:
    get:
      operationId: getMeeting
      parameters:
        - $ref: "#/components/parameters/MeetingId"
      responses:
        "200": { $ref: "#/components/responses/MeetingDocument" }
    put:
      operationId: putMeeting
      parameters:
        - $ref: "#/components/parameters/MeetingId"
      requestBody: { $ref: "#/components/requestBodies/MeetingUpdateRequest" }
      responses:
        "200": { $ref: "#/components/responses/MeetingDocument" }
    delete:
      operationId: deleteMeeting
      parameters:
        - $ref: "#/components/parameters/MeetingId"
      responses:
        "200": { $ref: "#/components/responses/DeletionDocument" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  requestBodies:
...
    MeetingStartRequest:
      required: true
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/MeetingStartDocument"
    MeetingUpdateRequest:
      required: true
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/MeetingUpdateDocument"
...
  responses:
...
    DeletionDocument:
      description: Deletion JSON:API document.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/DeletionDocument"
...
    MeetingCollectionDocument:
      description: Meeting JSON:API collection.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/MeetingCollectionDocument"
    MeetingDocument:
      description: Meeting JSON:API document.
      content:
        application/vnd.api+json:
          schema:
            $ref: "#/components/schemas/MeetingDocument"
...
  parameters:
...
    Status:
      name: status
      in: query
      schema: { type: string }
    TriggerSource:
      name: triggerSource
      in: query
      schema: { type: string }
    Tag:
      name: tag
      in: query
      schema: { type: string }
...
    ResearchTeamId:
      name: researchTeamId
      in: query
      schema: { type: string }
...
    PageLimit:
      name: page[limit]
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
    PageCursor:
      name: page[cursor]
      in: query
      schema:
        type: string
...
    MeetingId: { name: meetingId, in: path, required: true, schema: { type: string } }
...
  schemas:
...
    DeletionDocument:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/DeletionResource"
    DeletionResource:
      type: object
      required: [type, id, attributes]
      properties:
        type:
          type: string
          enum: [deletions]
        id:
          type: string
        attributes:
          type: object
          required: [status]
          properties:
            status:
              type: string
              enum: [deleted]
            unboundAccountIds:
              type: array
              items:
                type: integer
            unboundAccountCount:
              type: integer
...
    MeetingCollectionDocument:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/MeetingResource"
        links:
          $ref: "#/components/schemas/Links"
        meta:
          type: object
          additionalProperties: true
    MeetingDocument:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/MeetingResource"
    MeetingResource:
      type: object
      required: [type, id, attributes]
      properties:
        type: { type: string, enum: [meetings] }
        id: { type: string }
        attributes:
          $ref: "#/components/schemas/MeetingAttributes"
    MeetingAttributes:
      type: object
      properties:
        researchTeamId: { type: integer }
        topic: { type: string }
        status: { type: string }
        triggerSource: { type: string }
        summary:
          type: string
          nullable: true
        conclusion:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        recapStatus:
          type: string
          nullable: true
        recapUpdatedAt:
          type: string
          format: date-time
          nullable: true
        tokenBudget: { type: integer }
        runId:
          type: string
          nullable: true
        runAttempt: { type: integer }
        heartbeatAt:
          type: string
          format: date-time
          nullable: true
        autoRequeueCount: { type: integer }
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt: { type: string, format: date-time }
        trustReport:
          $ref: "#/components/schemas/MeetingTrustReport"
    MeetingTrustReport:
      type: object
      properties:
        generatedAt: { type: string, format: date-time }
        eventCount: { type: integer }
        evidenceCount: { type: integer }
        citationCount: { type: integer }
        modelSnapshotCount: { type: integer }
        promptSnapshotCount: { type: integer }
        confidence:
          type: string
          enum: [unknown, low, medium, high]
        evidence:
          type: array
          items:
            type: object
            additionalProperties: true
        citations:
          type: array
          items:
            type: object
            additionalProperties: true
        modelSnapshots:
          type: array
          items:
            type: object
            additionalProperties: true
        promptSnapshots:
          type: array
          items:
            type: object
            additionalProperties: true
        claimEvidenceBindings:
          type: array
          items:
            type: object
            additionalProperties: true
        evidenceGateStatus:
          type: string
          enum: [pass, warning, blocked]
        unsupportedClaimCount: { type: integer }
        unsupportedClaims:
          type: array
          items:
            $ref: "#/components/schemas/MeetingTrustUnsupportedClaim"
        evidenceGate:
          $ref: "#/components/schemas/MeetingTrustEvidenceGate"
        conclusionHistory:
          type: array
          items:
            type: object
            additionalProperties: true
        conclusionDiffs:
          type: array
          items:
            type: object
            additionalProperties: true
        sentenceReviews:
          type: array
          items:
            $ref: "#/components/schemas/MeetingTrustReviewAttributes"
        recapActionReviewStatus:
          type: string
          enum: [clear, needs_review, blocked]
        recapActionSuggestionCount: { type: integer }
        recapActionSuggestions:
          type: array
          items:
            $ref: "#/components/schemas/MeetingRecapActionSuggestion"
        recapActionReviewCount: { type: integer }
        recapActionReviews:
          type: array
          items:
            $ref: "#/components/schemas/MeetingRecapActionReviewAttributes"
        claimBreakdown:
          type: object
          properties:
            facts:
              type: array
              items: { type: string }
            assumptions:
              type: array
              items: { type: string }
            inferences:
              type: array
              items: { type: string }
        evidenceGaps:
          type: array
          items: { type: string }
        gaps:
          type: array
          items: { type: string }
    MeetingTrustEvidenceGate:
      type: object
      properties:
        status:
          type: string
          enum: [pass, warning, blocked]
        policy: { type: string }
        blockingClaimTypes:
          type: array
          items: { type: string }
        checkedClaimCount: { type: integer }
        supportedClaimCount: { type: integer }
        unsupportedClaimCount: { type: integer }
        unsupportedClaims:
          type: array
          items:
            $ref: "#/components/schemas/MeetingTrustUnsupportedClaim"
    MeetingTrustUnsupportedClaim:
      type: object
      properties:
        id: { type: string }
        eventId: { type: integer }
        sequence: { type: integer }
        roleKey: { type: string }
        claimType: { type: string }
        claim: { type: string }
        evidenceEventIds:
          type: array
          items: { type: integer }
        citationIds:
          type: array
          items: { type: string }
        reason: { type: string }
        createdAt: { type: string, format: date-time }
    MeetingRecapActionSuggestion:
      type: object
      properties:
        id: { type: string }
        eventId: { type: integer }
        sequence: { type: integer }
        roleKey: { type: string }
        status: { type: string }
        policy: { type: string }
        disposition: { type: string }
        reason: { type: string }
        actionType: { type: string }
        actionIndex: { type: integer }
        spec:
          type: object
          additionalProperties: true
        evidenceSummary:
          type: object
          additionalProperties: true
        latestReview:
          $ref: "#/components/schemas/MeetingRecapActionReviewAttributes"
        createdAt: { type: string, format: date-time }
...
    MeetingTrustReviewAttributes:
      type: object
      properties:
        meetingId: { type: integer }
        eventId: { type: integer }
        sequence: { type: integer }
        sentenceId: { type: string }
        sentence: { type: string }
        verdict:
          type: string
          enum: [confirmed, needs_evidence, rejected, superseded]
        citationIds:
          type: array
          items: { type: string }
        evidenceEventIds:
          type: array
          items: { type: integer }
        comment:
          type: string
          nullable: true
        reviewer: { type: string }
        reviewedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    MeetingRecapActionReviewAttributes:
      type: object
      properties:
        meetingId: { type: integer }
        eventId: { type: integer }
        sequence: { type: integer }
        suggestionId: { type: string }
        sourceEventId: { type: integer }
        sourceSequence: { type: integer }
        actionIndex: { type: integer }
        actionType: { type: string }
        decision:
          type: string
          enum: [approved, needs_evidence, rejected, superseded]
        executionDisposition: { type: string }
        citationIds:
          type: array
          items: { type: string }
        evidenceEventIds:
          type: array
          items: { type: integer }
        comment:
          type: string
          nullable: true
        reviewer: { type: string }
        reviewedAt:
          type: string
          format: date-time
        actionSpec:
          type: object
          additionalProperties: true
        originalStatus: { type: string }
        originalPolicy: { type: string }
        originalDisposition: { type: string }
        originalReason: { type: string }
        createdAt:
          type: string
          format: date-time
...
    MeetingStartDocument:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [type, attributes]
          properties:
            type: { type: string, enum: [meetings] }
            attributes:
              type: object
              required: [researchTeamId, topic]
              properties:
                researchTeamId: { type: integer }
                topic: { type: string }
                triggerSource: { type: string }
                context:
                  type: object
                  additionalProperties: true
    MeetingUpdateDocument:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [type, attributes]
          properties:
            type: { type: string, enum: [meetings] }
            id: { type: string }
            attributes:
              type: object
              properties:
                topic: { type: string }
                triggerSource: { type: string }
                tags:
                  type: array
                  items: { type: string }
...
    Links:
      type: object
      properties:
        self:
          type: string
        next:
          type: string
        prev:
          type: string